Moves FormFieldImageVideoLinkDocumentBrowsing
